test(benchmark): split W8A8 varlen attention benchmark
