+++ tools/select_tests.py
@@ -123,9 +123,6 @@ EXPLICIT_SOURCE_TO_TESTS = {
 
 # Same for benchmarks: keep explicit entries only for non-standard names that cannot be inferred from the source stem.
 EXPLICIT_SOURCE_TO_BENCHMARKS = {
-    "src/flaggems_vllm/ops/flash_attn_varlen_func_w8a8_fp8.py": [
-        "benchmark/test_flash_attn_varlen_func.py"
-    ],
     "src/flaggems_vllm/ops/rotary_embedding.py": [
         "benchmark/test_apply_rotary_pos_emb.py"
     ],
